requirements.txt
pytest
scikit-learn
pandas

fastapi 
uvicorn 
sqlalchemy
pypdf 
pydantic
joblib
argon2-cffi  
psycopg2-binary 
passlib 
python-multipart
python-dotenv
python-jose[cryptography]
bcrypt==4.2.0
alembic
pydantic[email]
email-validator
requests
xgboost